renamed RADD to CORE
